Add files via upload
Even or Odd Checker
Description
This C++ program is designed to determine whether a given number is even or odd. It provides a simple and efficient way to check the parity of any integer input. 🧮

How It Works
The program prompts the user to input a number, after which it performs a modulo operation to check whether the number is even or odd. If the input number is evenly divisible by 2, it is classified as even; otherwise, it is considered odd. 🔄